Endless pagination will no longer requery for ignored posts which weren't loaded

+++ app/src/main/java/com/lucasg234/protesttracker/mainactivity/ComposeFragment.java
@@ -242,8 +242,10 @@ public class ComposeFragment extends Fragment {
                 Log.i(TAG, "Saved post successfully");
                 mBinding.composeEditText.setText("");
                 mBinding.composeImagePreview.setImageBitmap(null);
-                // Delete the take image from internal storage
-                mInternalImageStorage.delete();
+                // Delete the take image from internal storage if there is one
+                if (mInternalImageStorage != null) {
+                    mInternalImageStorage.delete();
+                }
                 mInternalImageStorage = null;
             }
         });
